DXGL 0.0.5: Add more mode scaling options Add GetDC() GDI access to surfaces Add Random Noise test Add GDI Test patterns Add Random GDI object test
git-svn-id: https://www.williamfeely.info/svn/dxgl@5 8a90861a-4eca-46d5-b744-240ff16d0c4d

+++ cfgmgr/crc32.cpp
@@ -51,7 +51,7 @@ int main( int argc, const char *argv[] )
 {
     FILE *file = NULL;
     const char *filename;
-    unsigned long argIdx;
+    int argIdx;
     unsigned long crc32;
     int err;
